CHRONAM-2403: Batch delete working. Add additional decrementer line for failed messages to make counter continue to run in dynamo
Former-commit-id: 5080588285cb98a32342f51603d09b8298d8ff57 [formerly dee43891a2e0334f5c4dd9b2c6b52df992190e27]
Former-commit-id: 0f2ad08f9bf373e4bb90c0f6cab9bb0d1e631086 [formerly b434f81d0bc7a25e9920c44e40465d1a9b2aff6c]
Former-commit-id: 969bb8c199e2b48de4c7c3f0378d8f278d7b3b2e
Former-commit-id: c0b700260de5f635e285482a187a1860a5fb3853

+++ packages/cli/ndnp_openocr/cli.py
@@ -209,30 +209,33 @@ def delete(ctx, job_id, output_dir):
         print("[red]No output directory provided or found. Exiting.")
         return
 
-    bucket_name = ctx.obj["OUTPUT_BUCKET_NAME"]
+    # Confirmation prompt before deleting
+    if click.confirm(
+        f"Are you sure you want to delete S3 directory for job_id: {job_id} and local directory: {output_dir}?"
+    ):
+        bucket_name = ctx.obj["OUTPUT_BUCKET_NAME"]
 
-    # Delete S3 directory
-    s3 = boto3.resource("s3")
-    bucket = s3.Bucket(bucket_name)
-    prefix = f"{job_id}/"  # Assuming the job_id is the prefix for the files in S3
-    try:
-        bucket.objects.filter(Prefix=prefix).delete()
-        print(f"[green]Successfully deleted S3 directory for job_id: {job_id}")
-    except boto3.ClientError as e:
-        print(f"[red]Failed to delete S3 directory for job_id: {job_id}. Error: {e}")
+        # Delete S3 directory
+        s3 = boto3.resource("s3")
+        bucket = s3.Bucket(bucket_name)
+        prefix = f"{job_id}/"  # Assuming the job_id is the prefix for the files in S3
+        try:
+            bucket.objects.filter(Prefix=prefix).delete()
+            print(f"[green]Successfully deleted S3 directory for job_id: {job_id}")
+        except boto3.ClientError as e:
+            print(
+                f"[red]Failed to delete S3 directory for job_id: {job_id}. Error: {e}"
+            )
 
-    # Delete local directory
-    try:
-        shutil.rmtree(output_dir)
-        print(f"[green]Successfully deleted local directory: {output_dir}")
-    except OSError as e:
-        print(f"[red]Failed to delete local directory: {output_dir}. Error: {e}")
+        # Delete local directory
+        try:
+            shutil.rmtree(output_dir)
+            print(f"[green]Successfully deleted local directory: {output_dir}")
+        except OSError as e:
+            print(f"[red]Failed to delete local directory: {output_dir}. Error: {e}")
 
-    # Optionally, clear stored values if they were used
-    if not ctx.params["job_id"]:
-        keyring.delete_password("ndnp_openocr", "job_id")
-    if not ctx.params["output_dir"]:
-        keyring.delete_password("ndnp_openocr", "output_dir")
+    else:
+        print("[yellow]Deletion canceled.")
 
 
 @click.command(name="job_info")
@@ -275,6 +278,7 @@ cli.add_command(sync)
 cli.add_command(reprocess)
 cli.add_command(get)
 cli.add_command(job_info)
+cli.add_command(delete)
 
 if __name__ == "__main__":
     cli()
